Print to standard output, errors to standard error
* Print to standard output by default

If --output is not specified, psc-format will print to standard
output by default.

* Write errors to stdout and then exit with an error code
* Ensure standard input/error are using utf8
